move view extensions
