+++ app/src/main/java/mozilla/lockbox/extensions/view/ListPopupWindowItemClickObservable.kt
@@ -4,10 +4,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
 @file:JvmName("RxListPopupWindow")
 @file:JvmMultifileClass
 
-package mozilla.lockbox.rx
+package mozilla.lockbox.extensions.view
 
 import android.support.annotation.CheckResult
 import android.view.View
